Nu får man upp notis när tiden är ute, och i framtiden kommer man kunna ladda upp filer också

+++ components/file-upload.tsx
@@ -1,4 +1,5 @@
 "use client";
+
 import { OurFileRouter } from "@/app/api/uploadthing/core";
 import { UploadDropzone } from "@uploadthing/react";
 import { Trash } from "lucide-react";
@@ -9,30 +10,43 @@ import { Button } from "./ui/button";
 import { useToast } from "./ui/use-toast";
 
 interface ImageUploadProps {
-  onChange?: any;
-  onRemove: (value: UploadFileResponse[]) => void;
-  value: UploadFileResponse[];
+  onChange?: (value: UploadFileResponse[]) => void;
+  onRemove?: (value: UploadFileResponse[]) => void;
+  value?: UploadFileResponse[];
 }
 
 export default function FileUpload({
   onChange,
   onRemove,
-  value,
+  value = [],
 }: ImageUploadProps) {
   const { toast } = useToast();
+
   const onDeleteFile = (key: string) => {
-    const files = value;
-    let filteredFiles = files.filter((item) => item.key !== key);
-    onRemove(filteredFiles);
+    const filteredFiles = value.filter((item) => item.key !== key);
+    onRemove && onRemove(filteredFiles);
   };
+
   const onUpdateFile = (newFiles: UploadFileResponse[]) => {
-    onChange([...value, ...newFiles]);
+    onChange && onChange([...value, ...newFiles]);
   };
+
+  const allowedContent = ({ isUploading }: { isUploading: boolean }) => {
+    if (isUploading)
+      return (
+        <>
+          <p className="mt-2 text-sm text-slate-400 animate-pulse">
+            Img Uploading...
+          </p>
+        </>
+      );
+  };
+
   return (
-    <div>
+    <div style={{ width: '31rem' }}>
       <div className="mb-4 flex items-center gap-4">
         {!!value.length &&
-          value?.map((item) => (
+          value.map((item) => (
             <div
               key={item.key}
               className="relative w-[200px] h-[200px] rounded-md overflow-hidden"
@@ -64,20 +78,8 @@ export default function FileUpload({
             className="dark:bg-zinc-800 py-2 ut-label:text-sm ut-allowed-content:ut-uploading:text-red-300"
             endpoint="imageUploader"
             config={{ mode: "auto" }}
-            content={{
-              allowedContent({ isUploading }) {
-                if (isUploading)
-                  return (
-                    <>
-                      <p className="mt-2 text-sm text-slate-400 animate-pulse">
-                        Img Uploading...
-                      </p>
-                    </>
-                  );
-              },
-            }}
+            content={{ allowedContent }}
             onClientUploadComplete={(res) => {
-              // Do something with the response
               const data: UploadFileResponse[] | undefined = res;
               if (data) {
                 onUpdateFile(data);
